Feature/updated ERM (#34)
* updated ERM

* removed entity SeatingPlan

* set id to datatype long

* filmShow saves movie & Methods implemented

* Noch zwei kleine Verbesserungsvorschläge (#35)

* upate

* fix test

+++ src/main/java/com/kinoticket/backend/repositories/SeatsRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SeatsRepository extends JpaRepository<Seat, Integer> {
+public interface SeatsRepository extends JpaRepository<Seat, Long> {
 }
